+++ crates/forge/tests/e2e/running.rs
@@ -26,11 +26,11 @@ pub fn setup_package(package_name: &str) -> TempDir {
                 [package]
                 name = "{}"
                 version = "0.1.0"
-        
+
                 [[target.starknet-contract]]
                 sierra = true
                 casm = true
-        
+
                 [dependencies]
                 starknet = "2.1.0-rc2"
                 cheatcodes = {{ path = "{}" }}
@@ -53,30 +53,30 @@ fn simple_package() {
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 11 test(s) and 5 test file(s)
-        Running 1 test(s) from src/lib.cairo
-        [PASS] src::test_fib
+        Running 1 test(s) from simple_package package
+        [PASS] simple_package::test_fib
         Running 1 test(s) from tests/contract.cairo
-        [PASS] contract::contract::call_and_invoke
+        [PASS] contract::call_and_invoke
         Running 2 test(s) from tests/ext_function_test.cairo
-        [PASS] ext_function_test::ext_function_test::test_my_test
-        [PASS] ext_function_test::ext_function_test::test_simple
+        [PASS] ext_function_test::test_my_test
+        [PASS] ext_function_test::test_simple
         Running 6 test(s) from tests/test_simple.cairo
-        [PASS] test_simple::test_simple::test_simple
-        [PASS] test_simple::test_simple::test_simple2
-        [PASS] test_simple::test_simple::test_two
-        [PASS] test_simple::test_simple::test_two_and_two
-        [FAIL] test_simple::test_simple::test_failing
-        
+        [PASS] test_simple::test_simple
+        [PASS] test_simple::test_simple2
+        [PASS] test_simple::test_two
+        [PASS] test_simple::test_two_and_two
+        [FAIL] test_simple::test_failing
+
         Failure data:
             original value: [8111420071579136082810415440747], converted to a string: [failing check]
-        
-        [FAIL] test_simple::test_simple::test_another_failing
-        
+
+        [FAIL] test_simple::test_another_failing
+
         Failure data:
             original value: [8111420071579136082810415440747], converted to a string: [failing check]
-        
+
         Running 1 test(s) from tests/without_prefix.cairo
-        [PASS] without_prefix::without_prefix::five
+        [PASS] without_prefix::five
         Tests: 9 passed, 2 failed, 0 skipped
         "#});
 }
@@ -95,11 +95,11 @@ fn simple_package_with_git_dependency() {
             [package]
             name = "simple_package"
             version = "0.1.0"
-    
+
             [[target.starknet-contract]]
             sierra = true
             casm = true
-    
+
             [dependencies]
             starknet = "2.1.0-rc2"
             cheatcodes = { git = "https://github.com/foundry-rs/starknet-foundry.git" }
@@ -114,30 +114,30 @@ fn simple_package_with_git_dependency() {
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 11 test(s) and 5 test file(s)
-        Running 1 test(s) from src/lib.cairo
-        [PASS] src::test_fib
+        Running 1 test(s) from simple_package package
+        [PASS] simple_package::test_fib
         Running 1 test(s) from tests/contract.cairo
-        [PASS] contract::contract::call_and_invoke
+        [PASS] contract::call_and_invoke
         Running 2 test(s) from tests/ext_function_test.cairo
-        [PASS] ext_function_test::ext_function_test::test_my_test
-        [PASS] ext_function_test::ext_function_test::test_simple
+        [PASS] ext_function_test::test_my_test
+        [PASS] ext_function_test::test_simple
         Running 6 test(s) from tests/test_simple.cairo
-        [PASS] test_simple::test_simple::test_simple
-        [PASS] test_simple::test_simple::test_simple2
-        [PASS] test_simple::test_simple::test_two
-        [PASS] test_simple::test_simple::test_two_and_two
-        [FAIL] test_simple::test_simple::test_failing
+        [PASS] test_simple::test_simple
+        [PASS] test_simple::test_simple2
+        [PASS] test_simple::test_two
+        [PASS] test_simple::test_two_and_two
+        [FAIL] test_simple::test_failing
 
         Failure data:
             original value: [8111420071579136082810415440747], converted to a string: [failing check]
 
-        [FAIL] test_simple::test_simple::test_another_failing
+        [FAIL] test_simple::test_another_failing
 
         Failure data:
             original value: [8111420071579136082810415440747], converted to a string: [failing check]
 
         Running 1 test(s) from tests/without_prefix.cairo
-        [PASS] without_prefix::without_prefix::five
+        [PASS] without_prefix::five
         Tests: 9 passed, 2 failed, 0 skipped
         "#});
 }
@@ -174,12 +174,12 @@ fn with_filter() {
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 2 test(s) and 5 test file(s)
-        Running 0 test(s) from src/lib.cairo
+        Running 0 test(s) from simple_package package
         Running 0 test(s) from tests/contract.cairo
         Running 0 test(s) from tests/ext_function_test.cairo
         Running 2 test(s) from tests/test_simple.cairo
-        [PASS] test_simple::test_simple::test_two
-        [PASS] test_simple::test_simple::test_two_and_two
+        [PASS] test_simple::test_two
+        [PASS] test_simple::test_two_and_two
         Running 0 test(s) from tests/without_prefix.cairo
         Tests: 2 passed, 0 failed, 0 skipped
         "#});
@@ -192,16 +192,16 @@ fn with_exact_filter() {
 
     snapbox
         .current_dir(&temp)
-        .arg("test_simple::test_simple::test_two")
+        .arg("test_simple::test_two")
         .arg("--exact")
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 1 test(s) and 5 test file(s)
-        Running 0 test(s) from src/lib.cairo
+        Running 0 test(s) from simple_package package
         Running 0 test(s) from tests/contract.cairo
         Running 0 test(s) from tests/ext_function_test.cairo
         Running 1 test(s) from tests/test_simple.cairo
-        [PASS] test_simple::test_simple::test_two
+        [PASS] test_simple::test_two
         Running 0 test(s) from tests/without_prefix.cairo
         Tests: 1 passed, 0 failed, 0 skipped
         "#});
@@ -218,7 +218,7 @@ fn with_non_matching_filter() {
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 0 test(s) and 5 test file(s)
-        Running 0 test(s) from src/lib.cairo
+        Running 0 test(s) from simple_package package
         Running 0 test(s) from tests/contract.cairo
         Running 0 test(s) from tests/ext_function_test.cairo
         Running 0 test(s) from tests/test_simple.cairo
@@ -237,7 +237,7 @@ fn with_print() {
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 1 test(s) and 2 test file(s)
-        Running 0 test(s) from src/lib.cairo
+        Running 0 test(s) from print_test package
         Running 1 test(s) from tests/test_print.cairo
         original value: [123], converted to a string: [{]
         original value: [6381921], converted to a string: [aaa]
@@ -246,7 +246,7 @@ fn with_print() {
         original value: [124], converted to a string: [|]
         original value: [149]
         original value: [439721161573], converted to a string: [false]
-        [PASS] test_print::test_print::test_print
+        [PASS] test_print::test_print
         Tests: 1 passed, 0 failed, 0 skipped
         "#});
 }
@@ -261,10 +261,10 @@ fn with_panic_data_decoding() {
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 4 test(s) and 2 test file(s)
-        Running 0 test(s) from src/lib.cairo
+        Running 0 test(s) from panic_decoding package
         Running 4 test(s) from tests/test_panic_decoding.cairo
-        [PASS] test_panic_decoding::test_panic_decoding::test_simple
-        [FAIL] test_panic_decoding::test_panic_decoding::test_panic_decoding
+        [PASS] test_panic_decoding::test_simple
+        [FAIL] test_panic_decoding::test_panic_decoding
 
         Failure data:
             original value: [123], converted to a string: [{]
@@ -274,12 +274,12 @@ fn with_panic_data_decoding() {
             original value: [124], converted to a string: [|]
             original value: [149]
 
-        [FAIL] test_panic_decoding::test_panic_decoding::test_panic_decoding2
+        [FAIL] test_panic_decoding::test_panic_decoding2
 
         Failure data:
             original value: [128]
 
-        [PASS] test_panic_decoding::test_panic_decoding::test_simple2
+        [PASS] test_panic_decoding::test_simple2
         Tests: 2 passed, 2 failed, 0 skipped
         "#});
 }
@@ -295,8 +295,6 @@ fn with_exit_first() {
             name = "simple_package"
             version = "0.1.0"
 
-            # See more keys and their definitions at https://docs.swmansion.com/scarb/docs/reference/manifest
-
             [dependencies]
             starknet = "2.1.0-rc2"
             cheatcodes = {{ path = "{}" }}
@@ -304,10 +302,10 @@ fn with_exit_first() {
             [[target.starknet-contract]]
             sierra = true
             casm = true
-            
+
             [tool.snforge]
             exit_first = true
-            "#, 
+            "#,
             Utf8PathBuf::from_str("../../cheatcodes")
                 .unwrap()
                 .canonicalize_utf8()
@@ -322,25 +320,25 @@ fn with_exit_first() {
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 11 test(s) and 5 test file(s)
-        Running 1 test(s) from src/lib.cairo
-        [PASS] src::test_fib
+        Running 1 test(s) from simple_package package
+        [PASS] simple_package::test_fib
         Running 1 test(s) from tests/contract.cairo
-        [PASS] contract::contract::call_and_invoke
+        [PASS] contract::call_and_invoke
         Running 2 test(s) from tests/ext_function_test.cairo
-        [PASS] ext_function_test::ext_function_test::test_my_test
-        [PASS] ext_function_test::ext_function_test::test_simple
+        [PASS] ext_function_test::test_my_test
+        [PASS] ext_function_test::test_simple
         Running 6 test(s) from tests/test_simple.cairo
-        [PASS] test_simple::test_simple::test_simple
-        [PASS] test_simple::test_simple::test_simple2
-        [PASS] test_simple::test_simple::test_two
-        [PASS] test_simple::test_simple::test_two_and_two
-        [FAIL] test_simple::test_simple::test_failing
+        [PASS] test_simple::test_simple
+        [PASS] test_simple::test_simple2
+        [PASS] test_simple::test_two
+        [PASS] test_simple::test_two_and_two
+        [FAIL] test_simple::test_failing
 
         Failure data:
             original value: [8111420071579136082810415440747], converted to a string: [failing check]
 
-        [SKIP] test_simple::test_simple::test_another_failing
-        [SKIP] without_prefix::without_prefix::five
+        [SKIP] test_simple::test_another_failing
+        [SKIP] without_prefix::five
         Tests: 8 passed, 1 failed, 2 skipped
         "#});
 }
@@ -355,25 +353,25 @@ fn with_exit_first_flag() {
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 11 test(s) and 5 test file(s)
-        Running 1 test(s) from src/lib.cairo
-        [PASS] src::test_fib
+        Running 1 test(s) from simple_package package
+        [PASS] simple_package::test_fib
         Running 1 test(s) from tests/contract.cairo
-        [PASS] contract::contract::call_and_invoke
+        [PASS] contract::call_and_invoke
         Running 2 test(s) from tests/ext_function_test.cairo
-        [PASS] ext_function_test::ext_function_test::test_my_test
-        [PASS] ext_function_test::ext_function_test::test_simple
+        [PASS] ext_function_test::test_my_test
+        [PASS] ext_function_test::test_simple
         Running 6 test(s) from tests/test_simple.cairo
-        [PASS] test_simple::test_simple::test_simple
-        [PASS] test_simple::test_simple::test_simple2
-        [PASS] test_simple::test_simple::test_two
-        [PASS] test_simple::test_simple::test_two_and_two
-        [FAIL] test_simple::test_simple::test_failing
+        [PASS] test_simple::test_simple
+        [PASS] test_simple::test_simple2
+        [PASS] test_simple::test_two
+        [PASS] test_simple::test_two_and_two
+        [FAIL] test_simple::test_failing
 
         Failure data:
             original value: [8111420071579136082810415440747], converted to a string: [failing check]
 
-        [SKIP] test_simple::test_simple::test_another_failing
-        [SKIP] without_prefix::without_prefix::five
+        [SKIP] test_simple::test_another_failing
+        [SKIP] without_prefix::five
         Tests: 8 passed, 1 failed, 2 skipped
         "#});
 }
@@ -388,8 +386,6 @@ fn exit_first_flag_takes_precedence() {
             [package]
             name = "simple_package"
             version = "0.1.0"
-
-            # See more keys and their definitions at https://docs.swmansion.com/scarb/docs/reference/manifest
 
             [dependencies]
             starknet = "2.1.0-rc2"
@@ -412,25 +408,25 @@ fn exit_first_flag_takes_precedence() {
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 11 test(s) and 5 test file(s)
-        Running 1 test(s) from src/lib.cairo
-        [PASS] src::test_fib
+        Running 1 test(s) from simple_package package
+        [PASS] simple_package::test_fib
         Running 1 test(s) from tests/contract.cairo
-        [PASS] contract::contract::call_and_invoke
+        [PASS] contract::call_and_invoke
         Running 2 test(s) from tests/ext_function_test.cairo
-        [PASS] ext_function_test::ext_function_test::test_my_test
-        [PASS] ext_function_test::ext_function_test::test_simple
+        [PASS] ext_function_test::test_my_test
+        [PASS] ext_function_test::test_simple
         Running 6 test(s) from tests/test_simple.cairo
-        [PASS] test_simple::test_simple::test_simple
-        [PASS] test_simple::test_simple::test_simple2
-        [PASS] test_simple::test_simple::test_two
-        [PASS] test_simple::test_simple::test_two_and_two
-        [FAIL] test_simple::test_simple::test_failing
+        [PASS] test_simple::test_simple
+        [PASS] test_simple::test_simple2
+        [PASS] test_simple::test_two
+        [PASS] test_simple::test_two_and_two
+        [FAIL] test_simple::test_failing
 
         Failure data:
             original value: [8111420071579136082810415440747], converted to a string: [failing check]
 
-        [SKIP] test_simple::test_simple::test_another_failing
-        [SKIP] without_prefix::without_prefix::five
+        [SKIP] test_simple::test_another_failing
+        [SKIP] without_prefix::five
         Tests: 8 passed, 1 failed, 2 skipped
         "#});
 }
@@ -445,15 +441,15 @@ fn using_corelib_names() {
         .assert()
         .success()
         .stdout_matches(indoc! {r#"Collected 4 test(s) and 5 test file(s)
-        Running 0 test(s) from src/lib.cairo
+        Running 0 test(s) from using_corelib_names package
         Running 1 test(s) from tests/bits.cairo
-        [PASS] bits::bits::test_names
+        [PASS] bits::test_names
         Running 1 test(s) from tests/math.cairo
-        [PASS] math::math::test_names
+        [PASS] math::test_names
         Running 1 test(s) from tests/test.cairo
-        [PASS] test::test::test_names
+        [PASS] test::test_names
         Running 1 test(s) from tests/types.cairo
-        [PASS] types::types::test_names
+        [PASS] types::test_names
         Tests: 4 passed, 0 failed, 0 skipped
         "#});
 }
@@ -472,28 +468,28 @@ fn should_panic() {
         .success()
         .stdout_matches(indoc! { r#"
         Collected 6 test(s) and 2 test file(s)
-        Running 0 test(s) from src/lib.cairo
+        Running 0 test(s) from should_panic_test package
         Running 6 test(s) from tests/should_panic_test.cairo
-        [PASS] should_panic_test::should_panic_test::should_panic_no_data
+        [PASS] should_panic_test::should_panic_no_data
 
         Success data:
             original value: [0], converted to a string: []
 
-        [PASS] should_panic_test::should_panic_test::should_panic_check_data
-        [PASS] should_panic_test::should_panic_test::should_panic_multiple_messages
-        [FAIL] should_panic_test::should_panic_test::should_panic_with_non_matching_data
+        [PASS] should_panic_test::should_panic_check_data
+        [PASS] should_panic_test::should_panic_multiple_messages
+        [FAIL] should_panic_test::should_panic_with_non_matching_data
 
         Failure data:
             Incorrect panic data
             Actual:    [8111420071579136082810415440747] (failing check)
             Expected:  [0] ()
 
-        [FAIL] should_panic_test::should_panic_test::didnt_expect_panic
+        [FAIL] should_panic_test::didnt_expect_panic
 
         Failure data:
             original value: [156092886226808350968498952598218238307], converted to a string: [unexpected panic]
 
-        [FAIL] should_panic_test::should_panic_test::expected_panic_but_didnt
+        [FAIL] should_panic_test::expected_panic_but_didnt
         Tests: 3 passed, 3 failed, 0 skipped
         "#});
 }