+++ CHANGELOG.md
@@ -45,6 +45,12 @@ cheatcodes = { git = "https://github.com/foundry-rs/starknet-foundry", tag = "0.
 
 - Fails test when user tries to use syscalls not supported by forge test runner
 
+### Cast
+
+#### Added
+
+- Added `--class-hash` flag to account create/deploy, allowing for custom openzeppelin account contract class hash 
+
 ## [0.3.0] - 2023-08-02
 
 ### Forge